refactor(typings): move common dependencies to typings package
